[nextest-runner] move ExecutionResult and AbortStatus to reporter/events.rs
Even more thinning out of `runner/imp.rs`.
