+++ nextest-runner/tests/integration/fixtures.rs
@@ -17,9 +17,9 @@ use nextest_runner::{
         BinaryList, RustBuildMeta, RustTestArtifact, TestExecuteContext, TestList, TestListState,
     },
     platform::BuildPlatforms,
-    reporter::events::{ExecutionStatuses, RunStats, TestEventKind},
+    reporter::events::{AbortStatus, ExecutionResult, ExecutionStatuses, RunStats, TestEventKind},
     reuse_build::PathMapper,
-    runner::{configure_handle_inheritance, AbortStatus, ExecutionResult, TestRunner},
+    runner::{configure_handle_inheritance, TestRunner},
     target_runner::TargetRunner,
     test_filter::{FilterBound, TestFilterBuilder},
     test_output::{ChildExecutionOutput, ChildOutput},
